ftell utilise lseek. Pas encore testé.

+++ libc/include/stdio.h
@@ -244,7 +244,7 @@ void init_stdfiles(void);
 
 int fseek(FILE *stream, long offset, int whence);
 
-int ftell(FILE *stream);
+long ftell(FILE *stream);
 
 void rewind(FILE *stream);
